requirements.txt
flask
requests
fake